requirements.txt
pillow>=10.0.0
pytesseract>=0.3.10
pytest>=8.0.0
